Simplify patient-records view and localization.
I simplified the PatientRecordsPresenter. Since the
_patient_records_display partial is displaying records of many models,
there is going to be some aggregation code.  I don't want to put it in
the controller or the view. Could it go in the Record model
(form-object)? Probably. It's all display-related, though, so a
presenter seems like a decent option.

I was using the i18n_alchemy gem to parse and format date input and
output. I was using little of the gem's capabilities, rails already
handles the output part, and rolling my own input-parsing wasn't
difficult or much code.