Add rescuing of RecordNotFound exceptions.
A few cases where records may not be found were handled. I should be
handling all cases now.

TODO: If a RecordNotFound exception results from an Ajax request, then
the flash will be set but the page will NOT be refreshed. Therefore,
that flash will display on the next request. Fix that.