Simplify PatientsController.
I was manually deleting a patient's records before deleting a patient. I
removed that and added dependent: destroy to the associations in the
Patient mode. (The foreign key constraints in the database would delete
the associated records even without dependent: :destroy but using it keeps
the destroy callbacks being called on associated records.)

I cleaned up the names of the methods that parse "true" and "false"
strings to Ruby booleans. Again, this isn't necessary - the database
will parse the strings for me. But I'd rather have "false" be false in
case I add logic that deals with those values in the future.