Fix patients/edit.js.erb to replace correct HTML.
The js was replacing the HTML of div#patient-form-container instead of
div#form-panel. This meant that existing errors were not being cleared
(from trying to create a new patient) and an extra
div#patient-form-container was being nested.

I changed the view to replace the HTML of div#form-panel, fixing the
problem. Note that this is consistent with the edit.js.erb for users and
data points.