Fix edit-patient Ajax links to update DOM correctly.
1. If the user clicked an "edit patient" ajax link, then the
find-patient form would be removed. I fixed that by replacing the HTML
of the correct element, #patient-form-container instead of #form-panel.
    NOTE: this is now different than the various record-edit forms or the
    user-edit form because those don't have additional content in
    #form-panel.

2. If error messages were displayed on /patients and the user clicked an
"edit patient" ajax link, the new-patient form would be replaced with
an edit form but the error messages would still be present.
The #error-explanation element is now removed when the following actions
are taken via ajax:
    * an "edit patient" link is clicked
    * a patient-edit is canceled