Change styling of page header, form-error alert.
Decrease page-header font-size.

Let the alert with either the number of form errors or the errors on
the base model (as opposed to a field) push the form content down - it's
a pain and uglier to use absolute positioning.

NOTE: I think the right way to style the records page would be to make
the form fields and the display for each record type a row. Let the rows
stack to produce the current layout but then the two sides of the page,
form and display, would stay in sync.
This is going to lend itself to a nested model form or aggregating
multiple forms with JavaScript or something similar.
See http://railscasts.com/episodes/196-nested-model-form-part-1